backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pillow
python-multipart
aiofiles
torch
torchvision
transformers
sentencepiece
open_clip_torch
huggingface-hub
spacy
numpy==1.26.4
scikit-learn